Move ElementDescriptor to Pimcore\Model\Element namespace (#6267)
* Move ElementDescriptor to Pimcore\Model\Element namespace

* Move ElementDescriptor to Pimcore\Model\Element namespace

* Move ElementDescriptor to Pimcore\Model\Element namespace

* fixed issue with getCacheKey() + some coding style fixes

* updated gist
